requirements.txt
python-telegram-bot==13.15
transformers
torch
python-dotenv
flask